lab1 4.cpp changes

+++ C-Obj/Lab1/4.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main(){
     int n, i;
-    float sum=0, avg;
+    double sum=0, avg;
     cout<<"Too oruulna uu: ";
     cin>>n;
     for(i=1;i<n+1;i++){
